+++ CHANGELOG.markdown
@@ -2,7 +2,7 @@
 * :underscore option added
 * `Insensitive::KeyClashError` added for safer operations
 * Bug fix: Preserve default/default_proc on merge/replace/insensitive
-* Subtle behavioral change in Insensitive#replace when called with an ordinary Hash
+* Subtle behavioral change in `Insensitive#replace` when called with an ordinary Hash
 
 ### 0.2.1 / 2012/02/10
 * Bug fix: Insensitive `fetch`